Wrapper for LAPACK Routine DSYGVX
Adds the `generalized_symmetric_eigenproblem` module, which contains a wrapper for the LAPACK 
routine `DSYGVX` for solving real generalized symmetric-definite eigenproblems, as well as 
corresponding unit tests.

The module is written in such a way, that it can be extended, in the future, to include the 
routine `ZHEGVX` for treating the complex case.

The subroutine `solve_generalized_symmetric_eigenproblem` is designed to be used within the 
`exciting` routine `seceqnfv`, where the size of the arrays containing eigenvectors and 
eigenvalues is already given.